refactor(moyo-wasm): split lib.rs into per-domain modules
lib.rs grew to ~500 lines mixing analyze_cell, space-group and
layer-group bindings. Split into:

- common.rs     - MoyoOperation DTO, to_array9, map_moyo_err helpers
- dataset.rs    - MoyoCell/Lattice/AngleTolerance/MoyoDataset + analyze_cell
- space_group.rs - space-group lookups (Setting/Centering/HallSymbolEntry
                  /ArithmeticCrystalClass/SpaceGroupType + 4 wasm fns)
- layer_group.rs - layer-group lookups (same shape for the 80 LGs)

lib.rs is now just module declarations under the wasm32 cfg gate. Also
swap the local to_array3 helper for moyo::utils::to_3_slice (same
behaviour, already public from the core crate).
